chore: refactor all of pages to use domain structure
