+++ src/pages/acceptable-use/index.js
@@ -6,16 +6,16 @@ import {
 } from "gatsby"
 
 // Component Imports
-import Layout from "../components/layout"
+import Layout from "../../components/layout"
 
 // Style imports
-import styles from "./privacy-policy.module.scss"
+import styles from "./index.module.scss"
 
 const AcceptableUse = () => {
   const data = useStaticQuery(
     graphql`
     query {
-      contentfulPrivacyPolicy {
+      contentfulAcceptableUsePage {
         heroText
         bodyText {
           childMarkdownRemark {
@@ -29,12 +29,12 @@ const AcceptableUse = () => {
     <Layout>
       <div className={styles.container}>
         <div className={styles.header}>
-          <h1>{data.contentfulPrivacyPolicy.heroText}</h1>
+          <h1>{data.contentfulAcceptableUsePage.heroText}</h1>
         </div>
         <div className={styles.body}>
           <div
             dangerouslySetInnerHTML={{
-              __html: data.contentfulPrivacyPolicy.bodyText.childMarkdownRemark.html,
+              __html: data.contentfulAcceptableUsePage.bodyText.childMarkdownRemark.html,
             }}
           />
         </div>
