+++ src/pages/faq/index.module.scss
