+++ src/pages/index.js
@@ -9,12 +9,10 @@ import {
   graphql,
   useStaticQuery,
 } from "gatsby"
-// import Img from "gatsby-image"
 import Carousel from 'react-bootstrap/Carousel'
 
 // Component Imports
 import Layout from "../components/layout"
-// import HomeBanner from '../components/homeBanner'
 
 // Style Imports
 import styles from "./index.module.scss"
