+++ src/pages/merchandise/index.js
@@ -5,11 +5,11 @@ import {
 } from "gatsby"
 
 // Component Imports
-import Layout from "../components/layout"
-import DonateCTA from "../components/DonateCTA"
+import Layout from "../../components/layout"
+import DonateCTA from "../../components/DonateCTA"
 
 // Style Imports
-import merchStyles from "./merchandise.module.scss"
+import merchStyles from "./index.module.scss"
 
 const MerchandisePage = () => {
   const data = useStaticQuery(
