+++ src/pages/technology/index.module.scss
